requirements.txt
absl-py==1.0.0
arrow==1.2.2
astunparse==1.6.3
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cycler==0.11.0
Flask==2.0.3
flatbuffers==2.0
fonttools==4.31.2
fr==3.0
gast==0.5.3
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.6.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h5py==3.6.0
idna==3.3
imgkit==1.2.2
importlib-metadata==4.11.3
importlib-resources==5.4.0
itsdangerous==2.1.1
Jinja2==3.1.0
joblib==1.1.0
jsonschema==4.4.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.0
libclang==13.0.0
Markdown==3.3.6
MarkupSafe==2.1.1
matplotlib==3.5.1
numpy==1.22.3
oauthlib==3.2.0
opencv-python==4.5.5.62
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.11
pynput==1.7.6
pyobjc==8.4
pyobjc-core==8.4
pyobjc-framework-Accessibility==8.4
pyobjc-framework-Accounts==8.4
pyobjc-framework-AddressBook==8.4
pyobjc-framework-AdServices==8.4
pyobjc-framework-AdSupport==8.4
pyobjc-framework-AppleScriptKit==8.4
pyobjc-framework-AppleScriptObjC==8.4
pyobjc-framework-ApplicationServices==8.4
pyobjc-framework-AppTrackingTransparency==8.4
pyobjc-framework-AudioVideoBridging==8.4
pyobjc-framework-AuthenticationServices==8.4
pyobjc-framework-AutomaticAssessmentConfiguration==8.4
pyobjc-framework-Automator==8.4
pyobjc-framework-AVFoundation==8.4
pyobjc-framework-AVKit==8.4
pyobjc-framework-BusinessChat==8.4
pyobjc-framework-CalendarStore==8.4
pyobjc-framework-CallKit==8.4
pyobjc-framework-CFNetwork==8.4
pyobjc-framework-ClassKit==8.4
pyobjc-framework-CloudKit==8.4
pyobjc-framework-Cocoa==8.4
pyobjc-framework-Collaboration==8.4
pyobjc-framework-ColorSync==8.4
pyobjc-framework-Contacts==8.4
pyobjc-framework-ContactsUI==8.4
pyobjc-framework-CoreAudio==8.4
pyobjc-framework-CoreAudioKit==8.4
pyobjc-framework-CoreBluetooth==8.4
pyobjc-framework-CoreData==8.4
pyobjc-framework-CoreHaptics==8.4
pyobjc-framework-CoreLocation==8.4
pyobjc-framework-CoreMedia==8.4
pyobjc-framework-CoreMediaIO==8.4
pyobjc-framework-CoreMIDI==8.4
pyobjc-framework-CoreML==8.4
pyobjc-framework-CoreMotion==8.4
pyobjc-framework-CoreServices==8.4
pyobjc-framework-CoreSpotlight==8.4
pyobjc-framework-CoreText==8.4
pyobjc-framework-CoreWLAN==8.4
pyobjc-framework-CryptoTokenKit==8.4
pyobjc-framework-DataDetection==8.4
pyobjc-framework-DeviceCheck==8.4
pyobjc-framework-DictionaryServices==8.4
pyobjc-framework-DiscRecording==8.4
pyobjc-framework-DiscRecordingUI==8.4
pyobjc-framework-DiskArbitration==8.4
pyobjc-framework-DVDPlayback==8.4
pyobjc-framework-EventKit==8.4
pyobjc-framework-ExceptionHandling==8.4
pyobjc-framework-ExecutionPolicy==8.4
pyobjc-framework-ExternalAccessory==8.4
pyobjc-framework-FileProvider==8.4
pyobjc-framework-FileProviderUI==8.4
pyobjc-framework-FinderSync==8.4
pyobjc-framework-FSEvents==8.4
pyobjc-framework-GameCenter==8.4
pyobjc-framework-GameController==8.4
pyobjc-framework-GameKit==8.4
pyobjc-framework-GameplayKit==8.4
pyobjc-framework-ImageCaptureCore==8.4
pyobjc-framework-IMServicePlugIn==8.4
pyobjc-framework-InputMethodKit==8.4
pyobjc-framework-InstallerPlugins==8.4
pyobjc-framework-InstantMessage==8.4
pyobjc-framework-Intents==8.4
pyobjc-framework-IntentsUI==8.4
pyobjc-framework-IOSurface==8.4
pyobjc-framework-iTunesLibrary==8.4
pyobjc-framework-KernelManagement==8.4
pyobjc-framework-LatentSemanticMapping==8.4
pyobjc-framework-LaunchServices==8.4
pyobjc-framework-libdispatch==8.4
pyobjc-framework-LinkPresentation==8.4
pyobjc-framework-LocalAuthentication==8.4
pyobjc-framework-LocalAuthenticationEmbeddedUI==8.4
pyobjc-framework-MailKit==8.4
pyobjc-framework-MapKit==8.4
pyobjc-framework-MediaAccessibility==8.4
pyobjc-framework-MediaLibrary==8.4
pyobjc-framework-MediaPlayer==8.4
pyobjc-framework-MediaToolbox==8.4
pyobjc-framework-Metal==8.4
pyobjc-framework-MetalKit==8.4
pyobjc-framework-MetalPerformanceShaders==8.4
pyobjc-framework-MetalPerformanceShadersGraph==8.4
pyobjc-framework-MetricKit==8.4
pyobjc-framework-MLCompute==8.4
pyobjc-framework-ModelIO==8.4
pyobjc-framework-MultipeerConnectivity==8.4
pyobjc-framework-NaturalLanguage==8.4
pyobjc-framework-NetFS==8.4
pyobjc-framework-Network==8.4
pyobjc-framework-NetworkExtension==8.4
pyobjc-framework-NotificationCenter==8.4
pyobjc-framework-OpenDirectory==8.4
pyobjc-framework-OSAKit==8.4
pyobjc-framework-OSLog==8.4
pyobjc-framework-PassKit==8.4
pyobjc-framework-PencilKit==8.4
pyobjc-framework-Photos==8.4
pyobjc-framework-PhotosUI==8.4
pyobjc-framework-PreferencePanes==8.4
pyobjc-framework-PushKit==8.4
pyobjc-framework-Quartz==8.4
pyobjc-framework-QuickLookThumbnailing==8.4
pyobjc-framework-ReplayKit==8.4
pyobjc-framework-SafariServices==8.4
pyobjc-framework-SceneKit==8.4
pyobjc-framework-ScreenSaver==8.4
pyobjc-framework-ScreenTime==8.4
pyobjc-framework-ScriptingBridge==8.4
pyobjc-framework-SearchKit==8.4
pyobjc-framework-Security==8.4
pyobjc-framework-SecurityFoundation==8.4
pyobjc-framework-SecurityInterface==8.4
pyobjc-framework-ServiceManagement==8.4
pyobjc-framework-ShazamKit==8.4
pyobjc-framework-Social==8.4
pyobjc-framework-SoundAnalysis==8.4
pyobjc-framework-Speech==8.4
pyobjc-framework-SpriteKit==8.4
pyobjc-framework-StoreKit==8.4
pyobjc-framework-SyncServices==8.4
pyobjc-framework-SystemConfiguration==8.4
pyobjc-framework-SystemExtensions==8.4
pyobjc-framework-UniformTypeIdentifiers==8.4
pyobjc-framework-UserNotifications==8.4
pyobjc-framework-UserNotificationsUI==8.4
pyobjc-framework-VideoSubscriberAccount==8.4
pyobjc-framework-VideoToolbox==8.4
pyobjc-framework-Virtualization==8.4
pyobjc-framework-Vision==8.4
pyobjc-framework-WebKit==8.4
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-macos==2.8.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tqdm==4.63.0
tweepy==4.6.0
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.7.0
